Add RestrictedTopicEntity
